% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepackDiagnosticsFiles.R
\name{repackDiagnosticsFiles}
\alias{repackDiagnosticsFiles}
\title{Repack diagnostics files}
\usage{
repackDiagnosticsFiles(dataFolder)
}
\arguments{
\item{dataFolder}{folder where the exported zip files for the diagnostics are stored. 
Zip files containing results from multiple databases can be placed in the same
folder.}
}
\description{
The preMergeDiagnostics files creates one zip file per database with diagnostics for all cohort groups
(target, outcome,strata). However, in order to show diagnostics results from multiple databases, these
have to be re-combined per cohort group.
This function will unpack all zip files found in `dataFolder`, 
and copy any zip files found in subfolders of these zip files into a new folder 'repacked'.
}
